refactor: rename cliTools→cliFiles, move browser/cli configs to profiles/
- Rename option key cliTools → cliFiles to align with browserFiles
- Rename internal cliToolsConfig → cliFilesConfig, config name → @voxpelli/cli-files
- Move base-configs/browser.js and base-configs/cli.js into profiles/ to
  distinguish file-scoped profile overrides from universal rule sets
- Add profiles/**/* to package.json files array
