+++ eslint.config.js
@@ -1,5 +1,5 @@
 import { voxpelli } from './index.js';
 
 export default [
-  ...voxpelli({ cliTools: ['tools/**/*.js'] }),
+  ...voxpelli({ cliFiles: ['tools/**/*.js'] }),
 ];
